requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
black==23.1.0
blessed==1.20.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
croniter==1.3.8
dateutils==0.6.12
deepdiff==6.3.0
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.88.0
frozenlist==1.3.3
fsspec==2023.3.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
lightning==2.0.0
lightning-cloud==0.5.31
lightning-utilities==0.8.0
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.8.7
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.5
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.3.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.22.0
starsessions==1.3.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tomli==2.0.1
torch==1.13.1
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.6
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
